+++ crates/biome_service/src/file_handlers/vue.rs
@@ -24,9 +24,9 @@ use super::parse_lang_from_script_opening_tag;
 pub struct VueFileHandler;
 
 lazy_static! {
-    // https://regex101.com/r/E4n4hh/3
+    // https://regex101.com/r/E4n4hh/4
     pub static ref VUE_FENCE: Regex = Regex::new(
-        r#"(?ixms)(?<opening><script[^>]*>)\n(?P<script>(?U:.*))</script>"#
+        r#"(?ixs)(?<opening><script[^>]*>)\r?\n(?<script>(?U:.*))</script>"#
     )
     .unwrap();
 }
@@ -73,7 +73,9 @@ impl VueFileHandler {
             .captures(text)
             .and_then(|captures| {
                 match parse_lang_from_script_opening_tag(captures.name("opening")?.as_str()) {
-                    Language::JavaScript => None,
+                    Language::JavaScript => {
+                        Some(JsFileSource::js_module().with_embedding_kind(EmbeddingKind::Vue))
+                    }
                     Language::TypeScript { .. } => {
                         Some(JsFileSource::ts().with_embedding_kind(EmbeddingKind::Vue))
                     }